backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic-settings
python-multipart
minio
celery
redis
psycopg2-binary
websockets
Pillow